Fix compilation and test

+++ interfaces/IGovernanceManager.sol
@@ -8,27 +8,27 @@ interface IGovernanceManager {
         DataTypes.ProposalAction[] calldata actions
     ) external;
 
-    function vote(uint24 proposalId, DataTypes.Ballot ballot) external;
+    function vote(uint16 proposalId, DataTypes.Ballot ballot) external;
 
     function getVoteTotals(
-        uint24 proposalId
+        uint16 proposalId
     ) external view returns (DataTypes.VoteTotals memory);
 
-    function tallyVote(uint24 proposalId) external;
+    function tallyVote(uint16 proposalId) external;
 
     function getCurrentPercentages(
-        uint24 proposalId
+        uint16 proposalId
     ) external view returns (uint256 for_, uint256 against, uint256 abstain);
 
-    function executeProposal(uint24 proposalId) external;
+    function executeProposal(uint16 proposalId) external;
 
     function getBallot(
         address voter,
-        uint24 proposalId
+        uint16 proposalId
     ) external view returns (DataTypes.Ballot);
 
     function getProposal(
-        uint24 proposalId
+        uint16 proposalId
     ) external view returns (DataTypes.Proposal memory);
 
     function listActiveProposals()
